+++ packages/delir-core/src/engine/Task/ClipRenderTask.ts
@@ -9,8 +9,8 @@ import { RealParameterValues, RealParameterValueTypes } from '../Engine'
 import { compileTypeScript } from '../ExpressionSupport/ExpressionCompiler'
 import * as ExpressionContext from '../ExpressionSupport/ExpressionContext'
 import ExpressionVM from '../ExpressionSupport/ExpressionVM'
-import * as RendererFactory from '../renderer'
-import { IRenderer } from '../renderer/renderer-base'
+import * as RendererFactory from '../Renderer'
+import { IRenderer } from '../Renderer/RendererBase'
 import RenderRequest from '../RenderRequest'
 import EffectRenderTask from './EffectRenderTask'
