+++ packages/delir-core/src/engine/index.ts
@@ -1,4 +1,4 @@
-import * as Renderer from './renderer/index'
+import * as Renderer from './Renderer/index'
 
 export {expressionContextTypeDefinition} from './ExpressionSupport/ExpressionContext'
 export {default as Engine} from './Engine'
