+++ packages/delir-core/src/engine/renderer/Text/Text.ts
@@ -5,7 +5,7 @@ import Type from '../../../plugin-support/type-descriptor'
 import { TypeDescriptor } from '../../../plugin-support/type-descriptor'
 import PreRenderingRequest from '../../PreRenderingRequest'
 import RenderingRequest from '../../RenderRequest'
-import { IRenderer } from '../renderer-base'
+import { IRenderer } from '../RendererBase'
 
 import ColorRGBA from '../../../values/color-rgba'
